+++ test/weighted_hetero_node_test.cpp
@@ -92,10 +92,10 @@ TEST(WeightedHeteroNodeTest, RemoveAdjacent) {
     WeightedHeteroNode<int> node1;
     node1.setAdjacent(4, 1.5);
     node1.setAdjacent(5, 2.0);
-    EXPECT_EQ(node1.getAdjacents().size(), 2);
+    EXPECT_EQ(node1.getAdjacents().size(), static_cast<size_t>(2));
     node1.removeAdjacent(5);
     const auto& adjacents = node1.getAdjacents();
-    EXPECT_EQ(adjacents.size(), 1);
+    EXPECT_EQ(adjacents.size(), static_cast<size_t>(1));
     EXPECT_FALSE(adjacents.contains(5));
 }
